Moved utilityFunctions.kt
